requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.4.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
djangorestframework==3.12.4
drf-spectacular==0.21.1
idna==3.2
importlib-metadata==4.8.1
inflection==0.5.1
jsonschema==4.4.0
kombu==5.1.0
prompt-toolkit==3.0.20
psycopg2==2.9.3
PyJWT==1.7.1
pyrsistent==0.18.1
python-http-client==3.3.2
pytz==2021.1
PyYAML==6.0
redis==3.5.3
requests==2.26.0
sendgrid==6.8.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
twilio==6.63.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
